Correción de algunos detalles en la tarea informacion_oportunidad

+++ src/main/java/com/backendproject/springBackend/informacionoportunidad/repository/InformacionOportunidadRepository.java
@@ -1,4 +1,3 @@
-
 package com.backendproject.springBackend.informacionoportunidad.repository;
 
 import com.backendproject.springBackend.informacionoportunidad.model.InformacionOportunidad;
@@ -6,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface InformacionOportunidadRepository extends JpaRepository<InformacionOportunidad, Integer> {
+public interface InformacionOportunidadRepository extends JpaRepository<InformacionOportunidad, Long> {
 }
